--keyvani:🔧add config files and setup

+++ src/tasks/tasks.service.ts
@@ -1,10 +1,12 @@
-import { CreateTaskDto } from './dto/create-task.dto';
+import { CreateTaskDto, CreateTaskResponseDto } from './dto/create-task.dto';
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { GetTasksFilterDto } from './dto/get-tasks-filter.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
 import { TaskStatus } from './task-status.enum';
 import { Repository } from 'typeorm';
+import { User } from 'src/auth/user.entity';
+import { plainToInstance } from 'class-transformer';
 
 @Injectable()
 export class TasksService {
@@ -17,9 +19,10 @@ export class TasksService {
   //   return this.tasks;
   // }
 
-  async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
+  async getTasks(filterDto: GetTasksFilterDto, user: User): Promise<Task[]> {
     const { status, search } = filterDto;
     const query = this.taskRepository.createQueryBuilder('task');
+    query.where('task.userId = :userId', { userId: user.id });
 
     if (status) {
       query.andWhere('task.status = :status', { status });
@@ -51,33 +54,50 @@ export class TasksService {
   //   return tasks;
   // }
 
-  async getTaskById(id: number): Promise<Task> {
-    const found = await this.taskRepository.findOne({ where: { id } });
+  async getTaskById(id: number, user: User): Promise<Task> {
+    const found = await this.taskRepository.findOne({
+      where: { id, userId: user.id },
+    });
     if (!found) {
-      throw new NotFoundException(`Task with id ${id} not found`);
+      throw new NotFoundException(
+        `Task with id ${id} for this user ${user.username} not found`,
+      );
     }
     return found;
   }
 
-  async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
+  async createTask(
+    createTaskDto: CreateTaskDto,
+    user: User,
+  ): Promise<CreateTaskResponseDto> {
     const { title, description } = createTaskDto;
     const task = new Task();
     task.title = title;
     task.description = description;
     task.status = TaskStatus.OPEN;
+    task.user = user;
     await task.save();
-    return task;
+
+    return plainToInstance(CreateTaskResponseDto, task, {
+      excludeExtraneousValues: true,
+    });
   }
 
-  async deleteTask(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id);
+  async deleteTask(id: number, user: User): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id });
     if (result.affected === 0) {
-      throw new NotFoundException(`Don't found task with id ${id}`);
+      throw new NotFoundException(
+        `Don't found task with id ${id} for this user ${user.username}`,
+      );
     }
   }
 
-  async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-    const currentTask = await this.getTaskById(id);
+  async updateTaskStatus(
+    id: number,
+    status: TaskStatus,
+    user: User,
+  ): Promise<Task> {
+    const currentTask = await this.getTaskById(id, user);
     currentTask.status = status;
     await currentTask.save();
     return currentTask;
